% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getVisNet}
\alias{getVisNet}
\title{visNetwork Rendering}
\usage{
getVisNet(nets, title = NULL, pre_title = NULL)
}
\arguments{
\item{nets}{a list of igrpah network objects}

\item{title}{main title}

\item{pre_title}{title before the main title, as in "<pre-titile>: <title>"}
}
\value{
a visNetwork representing the given networks
}
\description{
This function receives a list of igrpah network objects, and optionally a title and a pre-title
and returns a visNetwork representing the given networks (using the visNetwork package)
}
\examples{

data = rbinom(300*20, 1, .2)
data = matrix(data,300,20)
res <- cluster_concordance(mat)
nets <- res[[1]]

getVisNet(nets, title, pre_title)
}
\keyword{graphs}
\keyword{list}
\keyword{networks}
\keyword{visNetwork}
