% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize_community}
\alias{normalize_community}
\title{Normalized Community Matrix}
\usage{
normalize_community(mats)
}
\arguments{
\item{mats}{list of NxN matrices.}
}
\value{
the corresponding normalized patient community concordance matrix.
}
\description{
This function receives a list of concordnce NxN matrices (a patient community)
and returns the corresponding normalized patient community concordance matrix.
}
\details{
It does so by summing the concordance matrices of the given community,
and normalizing offdiagonal entries by dividing each entry.
by the sqrt of the product of the associated diagonal entries,
(giving it a value in [0,1]).
}
\examples{

normalized_community(mats)
}
\keyword{list,}
\keyword{matrices}
