% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{similarity_matrix}
\alias{similarity_matrix}
\title{Similarity Matrix}
\usage{
similarity_matrix(
  mat,
  concords = vecs2concords(mat),
  simil_measure = "ARI",
  reactive_dom = NULL,
  incProg_amount = 1
)
}
\arguments{
\item{mat}{matrix}

\item{concords}{list of concordance matrices (defaults to vecs2concords(mat)) to be compared.}

\item{simil_measure}{string value, similarity function used, can be either 'ARI' (default) or 'Euclidean'.}

\item{reactive_dom}{reactive domain in case used in a shiny app (will be used by incProgress()).}

\item{inProg_amount}{real number in [0,1], the amount of status bar to fill up (default is 1), in case used in a shiny app (will be used by incProgress()).}
}
\value{
the corresponding NxN similarity matrix (pair-wise similarities) using the Adjusted Hubert-Arabie RAND index (ARI).
}
\description{
This function receives a matrix where each row is a binary vector (e.g. symptoms' presence vector)
and returns the corresponding NxN similarity matrix (pair-wise similarities) using the Adjusted Hubert-Arabie RAND index (ARI).
}
\examples{
mat <- matrix(sample(c(1,0), 16, replace=TRUE), 3, 3)
similarity_matrix(mat)
}
\keyword{comparisons}
\keyword{matrix}
\keyword{pair-wise}
\keyword{similarity}
