% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vec2concord}
\alias{vec2concord}
\title{Concordance Network}
\usage{
vec2concord(vec)
}
\arguments{
\item{vec}{vector.}
}
\value{
the corresponding NxN concordance matrix.
}
\description{
This function receives a vector and returns the corresponding NxN concordance matrix.
}
\examples{

vec <- sample(c(1,0), 4, replace=TRUE)
vec2concord(vec)
}
\keyword{matrix}
\keyword{to}
\keyword{vec}
