% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vecs2concords}
\alias{vecs2concords}
\title{Concordance Network}
\usage{
vecs2concords(mat)
}
\arguments{
\item{mat}{NxN matrix.}
}
\value{
the corresponding list of NxN concordance matrices.
}
\description{
This function receives an NXN matrix and returns the corresponding list of NxN concordance matrices.
}
\examples{

mat <- matrix(sample(c(1,0), 16, replace=TRUE), 3, 3)
vecs2concords(mat)
}
\keyword{matrices}
\keyword{to}
\keyword{vecs}
